Fix campaign slot creation form
