+++ back/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "02/09/2024 8:11:23 PM"
+    "x-generation-date": "02/12/2024 7:09:49 PM"
   },
   "x-strapi-config": {
     "path": "/documentation",
